Fix cmake installation

+++ install.sh
@@ -16,7 +16,7 @@ install_apt() {
 
   if [ "$distname" = "Ubuntu" ] && [ "$distver" = "16.04" ]; then
     sudo add-apt-repository -y ppa:deadsnakes/ppa
-    sudo add-apt-repository ppa:jonathonf/vim
+    sudo add-apt-repository -y ppa:jonathonf/vim
   fi
 
   sudo apt-get update
@@ -28,10 +28,12 @@ install_apt() {
     sudo apt-get install -y python3.6 clang-format-6.0
 
     # Install a newer CMake version
-    mkdir -p ~/.cache/brglng/dotfiles/cmake
-    wget -c https://github.com/Kitware/CMake/releases/download/v3.14.4/cmake-3.14.4-Linux-x86_64.sh -O ~/.cache/brglng/dotfiles/cmake/cmake-3.14.4-Linux-x86_64.sh
-    mkdir -p ~/.local
-    sh ~/.cache/brglng/dotfiles/cmake/cmake-3.14.4-Linux-x86_64.sh --prefix=$HOME/.local --exclude-subdir
+    if [ ! -e ~/.local/bin/cmake ]; then
+      mkdir -p ~/.cache/brglng/dotfiles/cmake
+      wget -c https://github.com/Kitware/CMake/releases/download/v3.14.4/cmake-3.14.4-Linux-x86_64.sh -O ~/.cache/brglng/dotfiles/cmake/cmake-3.14.4-Linux-x86_64.sh
+      mkdir -p ~/.local
+      sh ~/.cache/brglng/dotfiles/cmake/cmake-3.14.4-Linux-x86_64.sh --prefix=$HOME/.local --exclude-subdir
+    fi
   else
     sudo apt-get install -y clang-format-7
   fi
